+++ Magic/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # 7.9.11
 
- - Don't restore bound wands if they have been set as not bound (or disabled) in configs
+ - Some changes and improvements to unbinding wands
 
 # 7.9.10